% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations.r
\docType{data}
\name{.lucC_build_intervals}
\alias{.lucC_build_intervals}
\title{Build Intervals of Data with Raster}
\usage{
.lucC_build_intervals(first_ras = NULL, second_ras = NULL)
}
\arguments{
\item{first_ras}{matrix. An interval between two dates.}

\item{second_ras}{matrix. An interval between two dates.}
}
\value{
A list with value of interval for each data set
}
\description{
Provide an valide interval from data set input.
And return a list with two intervals.
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
