% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_predicates_raster.R
\docType{data}
\name{.lucC_check_intervals}
\alias{.lucC_check_intervals}
\title{Build Intervals of Data with Raster}
\usage{
.lucC_check_intervals(first_int = NULL, second_int = NULL)
}
\arguments{
\item{first_int}{Date. An interval between two dates.}

\item{second_int}{Date. An interval between two dates.}
}
\value{
A list with value of interval for each data set
}
\description{
Provide an valide interval from data set input.
And return a list with two intervals.
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
