% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_load_data.R
\name{example_points_ts}
\alias{example_points_ts}
\title{Example CSV with a time series set}
\description{
Dataset of points for using with sits_getdata and acquire time series from WTSS. With data from a particular region from Santa Carmem municipality, Mato Grosso, Brazil.
}
