% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_load_data.R
\name{example_ts}
\alias{example_ts}
\title{example file of time series set from sits}
\description{
Dataset of example of time series set from sits in format RData to run some functions this package. With data from a particular region from Santa Carmem municipality, Mato Grosso, Brazil.
}
