% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_load_data.R
\name{example_TWDTW}
\alias{example_TWDTW}
\alias{example_TWDTW}
\title{Example CSV File}
\description{
Dataset of example of a time series set already classified using TWDTW in CSV format to run some functions this package. With data from a particular region from Santa Carmem municipality, Mato Grosso, Brazil.

Dataset of example of a time series set already classified using TWDTW in RData to run some functions this package. With data from a particular region from Santa Carmem municipality, Mato Grosso, Brazil.
}
