% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_load_data.R
\docType{data}
\name{lucC_data_preparation}
\alias{lucC_data_preparation}
\title{Format data as tibble format}
\usage{
lucC_data_preparation(data_tb = NULL)
}
\arguments{
\item{data_tb}{Dataframe. A time series file.}
}
\value{
Tibble with columns 'longitude', 'latitude', 'start_date', 'end_date', 'label', 'id', 'index'
}
\description{
Provide a format pattern to input data and make operations with Allen's relations and predicates. And return a tibble with values input
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

# open a CSV example
file = "./inst/example_csv_file.csv"

# read a raw csv and turn into format lucC with
input_tb_raw_csv <- file \%>\%
  read.csv(sep = ",", header = TRUE) \%>\%
  lucC_data_preparation()
input_tb_raw_csv

new_tb <- lucC_data_preparation(input_tb)

}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
