% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_result_measures.R
\docType{data}
\name{lucC_extract_frequency}
\alias{lucC_extract_frequency}
\title{Categorical frequency from a list of data.frames}
\usage{
lucC_extract_frequency (data_mtx.list = NULL, cores_in_parallel = 1)
}
\arguments{
\item{data_mtx.list}{Matrix. A list of matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS for each block of RasterBrick}

\item{cores_in_parallel}{Integer. Number of cores to use, in other words the amount of child processes will be run simultaneously. Default is 1 core, that means no one parallelization.}
}
\value{
Data frame with statistical measures
}
\description{
Provide a data.frame with count of categorical variables from a list of data.frame derived of processing with many blocks of RasterBricks
}
\examples{
\dontrun{

lucC_extract_frequency(data_mtx.list = NULL, cores_in_parallel = 1)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
