% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_load_data.R
\docType{data}
\name{lucC_fromJSON}
\alias{lucC_fromJSON}
\title{Open JSON file}
\usage{
lucC_fromJSON (path_json_file = NULL)
}
\arguments{
\item{path_json_file}{Character. Name path and file to open JSON file data}
}
\value{
Open JSON format to file
}
\description{
Open the input data in JSON format in an user's directory
}
\examples{
\dontrun{
# Open a data example
library(lucC)

lucC_starting_point()

json_file = "./inst/extdata/patterns/example_TWDTW.json"

input_tb_json <- json_file \%>\%
  lucC_fromJSON()
input_tb_json

}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
