% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_interval.R
\docType{data}
\name{lucC_interval}
\alias{lucC_interval}
\title{lucC Intervals}
\usage{
lucC_interval(first_date , second_date)
}
\arguments{
\item{first_date}{Date. An date value like '2012-03-02'.}

\item{second_date}{Date. An date value like '2013-03-02'.}
}
\value{
Interval value with two dates
}
\description{
Provide a lucC_interval, an input pattern format used by Allen's relations. And return a value of interval
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

# create some examples of intervals
time1 <- lucC_interval("2011-09-01","2011-10-01")
time2 <- lucC_interval("2011-09-15","2011-11-01")


}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
