% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations_extra.r
\docType{data}
\name{lucC_merge}
\alias{lucC_merge}
\title{LucC Merge}
\usage{
lucC_merge(first_raster = NULL, second_raster = NULL)
}
\arguments{
\item{first_raster}{Matrix. An interval between two dates.}

\item{second_raster}{Matrix. An interval between two dates.}
}
\value{
Data set merged with two data sets that values are in the same interval
}
\description{
Merge two data set by different columns and rows. Because base::merge replace values in initial years
}
\examples{
\dontrun{

a <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Forest",
                     time_interval = c("2001-09-01","2002-09-01"),
                     relation_interval = "contains", label = label,
                     timeline = timeline)
a

b <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Pasture",
                     time_interval = c("2001-09-01","2002-09-01"),
                     relation_interval = "contains", label = label,
                     timeline = timeline)
b

# merge
c <- lucC_merge(first_raster = a, second_raster = b)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
