% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_create_blocks.R
\docType{data}
\name{lucC_merge_blocks}
\alias{lucC_merge_blocks}
\title{Create blocks from RasterBrick in accordance with number of cells}
\usage{
lucC_merge_blocks (path_open_GeoTIFFs = NULL)
}
\arguments{
\item{path_open_GeoTIFFs}{Character. Name a path folder to OPEN raster images data.}
}
\value{
RasterBrick Mosaic.
}
\description{
Merge GeoTIFF splitted into parts. \url{https://stackoverflow.com/questions/29784829/r-raster-package-split-image-into-multiples}
}
\examples{
\dontrun{

lucC_merge_blocks(path_open_GeoTIFFs = NULL)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
