% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_plot_results.R
\docType{data}
\name{lucC_plot_bar_events}
\alias{lucC_plot_bar_events}
\title{Plot Barplot Maps}
\usage{
lucC_plot_bar_events (data_mtx = NULL,
custom_palette = FALSE, RGB_color = NULL, pixel_resolution = 250,
relabel = FALSE, original_labels = NULL, new_labels = NULL,
legend_text = "Legend:", column_legend = 2,
side_by_side = FALSE)
}
\arguments{
\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{custom_palette}{Boolean. A TRUE or FALSE value. If TRUE, user will provide its own color palette setting! Default is FALSE}

\item{RGB_color}{Character. A vector with color names to map legend, for example, c("Green","Blue"). Default is setting scale_colour_hue}

\item{pixel_resolution}{Numeric. Is a spatial resolution of the pixel. Default is 250 meters considering MODIS 250 m. See more at \url{https://modis.gsfc.nasa.gov/about/specifications.php}.}

\item{relabel}{Boolean. A TRUE or FALSE value. If TRUE, user will provide its own legend text setting! Default is FALSE}

\item{original_labels}{Character. A vector with original labels from legend text, for example, c("Forest","Pasture").}

\item{new_labels}{Character. A vector with new labels to legend text, for example, c("Mature_Forest","Pasture1").}

\item{legend_text}{Character. A text legend to show in plot, such as "Land use transitions:". Default is "Legend:".}

\item{column_legend}{Integer. A number with the desired number of columns in legend.}

\item{side_by_side}{Boolean. Make bar of a barplot a side-by-side plot. Default is FALSE.}
}
\value{
Plot a barplot in Y axis in square kilometers (Area km^2) = (Number of pixel *(pixel_resolution*pixel_resolution))/(1000*1000)
}
\description{
Plot barplot over time
}
\examples{
\dontrun{

lucC_plot_bar_events(data_mtx = raster_mtx, custom_palette = TRUE,
RGB_color = c("green", "orange"), pixel_resolution = 232,
relabel = TRUE, original_labels = c("Forest", "Pasture1"),
new_labels = c("Forest", "Pasture"), legend_text = "Legend: ",
column_legend = 1, side_by_side = TRUE)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
