% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_plot_maps_events.r
\docType{data}
\name{lucC_plot_maps_events}
\alias{lucC_plot_maps_events}
\title{Plot Events over Maps with lucC}
\usage{
lucC_plot_maps_events (data_tb = NULL, EPSG_WGS84 = TRUE, 
custom_palette = FALSE, RGB_color = NULL, plot_ncol = NULL, 
shape_point = 0, colour_point = "black" , size_point= 1, 
relabel = FALSE, original_labels = NULL, new_labels = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{EPSG_WGS84}{Character. A reference coordinate system. If TRUE, the values of latitude and longitude alredy use this coordinate system, if FALSE, the data set need to be transformed}

\item{custom_palette}{Boolean. A TRUE or FALSE value. If TRUE, user will provide its own color palette setting! Default is FALSE}

\item{RGB_color}{Character. A vector with color names to map legend, for example, c("Green","Blue"). Default is setting scale_colour_hue}

\item{plot_ncol}{Numeric. A number of columns to show the maps}

\item{shape_point}{Numeric or Character. Is a shape point for events highlighted over map. Default is 0.  This includes different points symbols commonly used in R as "pch", such as numeric values like 0 to square, 1 to circle and 4 to cross shape. And also other characters can be used including ".", "+", "*", "-", "#".}

\item{colour_point}{Numeric. Is a colour for the shape point for events highlighted over map. Default is black}

\item{size_point}{Numeric. Is a size of the shape point around pixels in plot. Default is 1}

\item{relabel}{Boolean. A TRUE or FALSE value. If TRUE, user will provide its own legend text setting! Default is FALSE}

\item{original_labels}{Character. A vector with original labels from legend text, for example, c("Forest","Pasture").}

\item{new_labels}{Character. A vector with new labels to legend text, for example, c("Mature_Forest","Pasture1").}
}
\value{
Plot with input data as colored map
}
\description{
Plot map ggplot2 for all events discovered in input data
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

# open a JSON file example
file_json = "./inst/extdata/patterns/example_TWDTW.json"

# open file JSON
input_tb_raw_json <- file_json \%>\%
  lucC_fromJSON()
input_tb_raw_json

# plot maps input data
lucC_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE, 
custom_palette = TRUE, RGB_color = c("#FFB266", "#1b791f", 
"#929e6e", "#f5e7a1"))

# define interval
time_ex1 <- lucC_interval("2002-01-01", "2014-01-01")

# apply predicate occur
ts_occur1 <- lucC_predicate_holds(geo_objects = input_tb_raw_json, 
object_properties = "Forest", event_time_intervals = time_ex1)
ts_occur1

# events over input map
lucC_plot_maps_events(ts_occur1, EPSG_WGS84 = TRUE, 
custom_palette = TRUE, RGB_color = c("#FFB266", "#1b791f", 
"#929e6e", "#f5e7a1"), shape_point = 0, colour_point = "black", 
size_point = 2.3)


}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
