% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_predicates_raster.R
\docType{data}
\name{lucC_pred_holds}
\alias{lucC_pred_holds}
\title{Predicate Allen Holds}
\usage{
lucC_pred_holds (raster_obj = NULL, raster_class = NULL,
time_interval = c("2000-01-01", "2004-01-01"),
relation_interval = "contains", label = NULL, timeline = NULL)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{raster_class}{Character. Name of the class of interest, such as 'Forest', to research}

\item{time_interval}{Interval. A time interval to verify if class is over or not}

\item{relation_interval}{Character. If a location HOLDS during all time interval 'equals' or can be appear in any
times 'contains'. Default is 'contains'}

\item{label}{Character Vector. All labels of each value of pixel from classified raster}

\item{timeline}{Character. A list of all dates of classified raster, timeline}
}
\value{
Matrix with all states which holds during a time interval
}
\description{
Provide a predicate HOLDS which evaluates as true when a class \code{c_i},
e.g. 'Forest', holds uring the interval \code{t_i}. Return a matrix with values within
defined interval
}
\examples{
\dontrun{

a <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Pasture",
time_interval = c("2001-09-01","2005-09-01"), relation_interval = "contains",
label = label, timeline = timeline)
a

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
