% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_predicates_raster.R
\docType{data}
\name{lucC_pred_recur}
\alias{lucC_pred_recur}
\title{Predicate Allen Holds}
\usage{
lucC_pred_recur (raster_obj = NULL, raster_class = NULL,
time_interval1 = c("2001-01-01", "2001-01-01"),
time_interval2 = c("2002-01-01", "2005-01-01"),
label = NULL, timeline = NULL)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{raster_class}{Character. Name of the class of interest, such as 'Forest', to research}

\item{time_interval1}{Interval. A first time interval to verify if class is over or not}

\item{time_interval2}{Interval. A second and non-overlapped time interval to verify if class is over or not}

\item{label}{Character. All labels of each value of pixel from classified raster}

\item{timeline}{Character. A list of all dates of classified raster, timeline}
}
\value{
Matrix with all states which holds during a time interval
}
\description{
Provide a predicate RECUR which evaluates as true when a location holds
a class \code{c_i}, e.g. 'Forest', during two non-continuous distinct intervals
\code{t_i} and \code{t_j}. Return a matrix with values within defined interval
}
\examples{
\dontrun{

b <- lucC_pred_recur(raster_obj = rb_sits, raster_class = "Forest",
                     time_interval1 = c("2001-09-01","2001-09-01"),
                     time_interval2 = c("2002-09-01","2016-09-01"),
                     label = label, timeline = timeline))
b

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
