% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_predicates_holds.r
\docType{data}
\name{lucC_predicate_holds}
\alias{lucC_predicate_holds}
\title{Predicate Allen Holds}
\usage{
lucC_predicate_holds (locations = NULL, location_properties = NULL, 
time_intervals = lucC_interval("2000-01-01", "2004-01-01"))
}
\arguments{
\item{locations}{Tibble. A tibble with values longitude and latitude and other values}

\item{location_properties}{Character. Name of value present in a row of the tibble, such as 'Forest' or other value}

\item{time_intervals}{Interval. A interval of time to verify if location_properties is over or 
not in lucC_interval format. Given a tibble with values, will be asserts if that location_properties 
of locations holds during a time interval.}
}
\value{
Tibble with all events hold during a time interval
}
\description{
Provide a predicate of Allen's which asserts that a property holds 
during a time interval. Return a tibble with value within defined interval
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

data("example_TWDTW")

input_tb_json <- example_TWDTW \%>\%
  lucC_standard_date_events(month_year = "09", day_month = "01")
input_tb_json

# example of application
time_ex1 <- lucC_interval("2000-09-01", "2003-09-01")
time_ex1
time_ex2 <- lucC_interval("2003-09-01", "2010-09-01")
time_ex2

# location_properties
properties <- "Forest"

# example predicate holds
lucC_predicate_holds(locations = input_tb_json, 
location_properties = "Forest", time_intervals = time_ex1)

lucC_predicate_holds(locations = input_tb_json, 
location_properties = properties, time_intervals = time_ex2)

}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
