% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations.r
\docType{data}
\name{lucC_relation_contains}
\alias{lucC_relation_contains}
\title{Allen Relation Contains}
\usage{
lucC_relation_contains(first_raster = NULL, second_raster = NULL)
}
\arguments{
\item{first_raster}{Matrix. An interval between two dates.}

\item{second_raster}{Matrix. An interval between two dates.}
}
\value{
Data set with merge of two data sets
}
\description{
Provide an Allen's interval relation CONTAINS which start time interval of the
raster_1 must be (<) less than the start time interval from raster_2, and end time interval
of the raster_1 must be (>) greater than the end time interval from raster_2. See more at (ALLEN, J. F.
"Maintaining knowledge about temporal intervals". Communications of the ACM, v(26), 11,
1983, 832-843. DOI: \url{http://dx.doi.org/10.1145/182.358434})
}
\examples{
\dontrun{

a <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Forest",
                     time_interval = c("2001-09-01","2015-09-01"),
                     relation_interval = "equals", label = label,
                     timeline = timeline)
a

b <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Cerrado",
                     time_interval = c("2003-09-01","2011-09-01"),
                     relation_interval = "equals", label = label,
                     timeline = timeline)
b

# contains
c <- lucC_relation_contains(first_raster = a, second_raster = b)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
