% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations_extra.r
\docType{data}
\name{lucC_relation_follows}
\alias{lucC_relation_follows}
\title{Allen Relation Follows}
\usage{
lucC_relation_follows(first_raster = NULL, second_raster = NULL)
}
\arguments{
\item{first_raster}{matrix. An interval between two dates.}

\item{second_raster}{matrix. An interval between two dates.}
}
\value{
Data set with merge of two data sets
}
\description{
Provide an Allen's interval relation to classified raster data.
And return data set with raster data sets merged
}
\examples{
\dontrun{

a <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Forest",
                     time_interval = c("2001-09-01","2003-09-01"),
                     relation_interval = "equals", label = label,
                     timeline = timeline)
a

b <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Cerrado",
                     time_interval = c("2004-09-01","2007-09-01"),
                     relation_interval = "equals", label = label,
                     timeline = timeline)
b

# follows
c <- lucC_relation_follows(first_raster = a, second_raster = b)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
