% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations_extra.r
\docType{data}
\name{lucC_relation_occurs}
\alias{lucC_relation_occurs}
\alias{lucC_relation_same_interval}
\title{Allen Relation Precedes}
\usage{
lucC_relation_occurs(first_raster = NULL, second_raster = NULL)
}
\arguments{
\item{first_raster}{Matrix. An interval between two dates.}

\item{second_raster}{Matrix. An interval between two dates.}
}
\value{
Data set with merge of two data sets that values are in the same interval
}
\description{
Provide an interval relation OCCURS which asserts that classes
of two distinct data set occurs in the same interval
}
\examples{
\dontrun{

a <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Forest",
                     time_interval = c("2001-09-01","2002-09-01"),
                     relation_interval = "contains", label = label,
                     timeline = timeline)
a

b <- lucC_pred_holds(raster_obj = rb_sits, raster_class = "Pasture",
                     time_interval = c("2001-09-01","2002-09-01"),
                     relation_interval = "contains", label = label,
                     timeline = timeline)
b

# occurs
c <- lucC_relation_occurs(first_raster = a, second_raster = b)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
