% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_remove_columns.R
\docType{data}
\name{lucC_remove_columns}
\alias{lucC_remove_columns}
\title{Save tibble as Raster in Folder}
\usage{
lucC_remove_columns(data_mtx = NULL, name_columns = NULL)
}
\arguments{
\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{name_columns}{Character. Name of columns to remove from data set}
}
\value{
Matrix with without selected columns
}
\description{
Remove columns from a data_mtx
}
\examples{
\dontrun{

# save RasterBrick in path
lucC_create_RasterBrick (path_open_GeoTIFFs = "~/Desktop/raster", path_save_RasterBrick = "~/Desktop")

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
