% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_result_format.R
\docType{data}
\name{lucC_result_format}
\alias{lucC_result_format}
\title{Predicate Allen Holds}
\usage{
lucC_result_format (raster_df = NULL)
}
\arguments{
\item{raster_df}{Matrix or Dataframe. A data set from predicate operation}
}
\value{
Tibble with all states
}
\description{
Provide a predicate of Allen's which asserts that a class holds
during a time interval. Return a tibble with value within defined interval
}
\examples{
\dontrun{

library(lucC)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
