% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_result_measures.R
\docType{data}
\name{lucC_result_measures}
\alias{lucC_result_measures}
\title{Plot Sequence Maps with lucC Events}
\usage{
lucC_result_measures (data_mtx = NULL, pixel_resolution = 250)
}
\arguments{
\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{pixel_resolution}{Numeric. Is a spatial resolution of the pixel. Default is 250 meters considering MODIS 250 m. See more at \url{https://modis.gsfc.nasa.gov/about/specifications.php}.}
}
\value{
Data frame with measures computed
}
\description{
Provide a data.frame with main measures to resultd from LUC Calculus application such as: Area (km2),
Cumulative Sum (CumSum), Relative Frequency and Cumulative Relative Frequency
}
\examples{
\dontrun{

lucC_result_measures(data_mtx = NULL, pixel_resolution = NULL)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
