% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_save_raster_GeoTIFF.R
\docType{data}
\name{lucC_save_GeoTIFF}
\alias{lucC_save_GeoTIFF}
\title{Save tibble as Raster in Folder}
\usage{
lucC_save_GeoTIFF(raster_obj = NULL, data_mtx = NULL, path_raster_folder = NULL)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{data_mtx}{Matrix. A matrix with pixel replaced, can be obtined using lucC_update_raster()}

\item{path_raster_folder}{Character. Name a path folder to save raster images data. If  doesn't exist, a new directory is created}
}
\value{
Raster in geotiff format to open using SIG
}
\description{
Save new_raster reclassified in a diretory defined by user
}
\examples{
\dontrun{

# save rasters in folder
lucC_save_GeoTIFF (raster_obj = rb_sits, data_mtx = new_raster, path_raster_folder = "~/Desktop/raster")

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
