% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_save_raster_GeoTIFF.R
\docType{data}
\name{lucC_save_raster_result}
\alias{lucC_save_raster_result}
\title{Update a RasterBrick with result from predicates}
\usage{
lucC_save_raster_result(raster_obj = NULL, data_mtx = NULL,
timeline = NULL, label = NULL, path_raster_folder = NULL)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{timeline}{Character. A list of all dates of classified raster, timeline}

\item{label}{Character Vector. All labels of each value of pixel from classified raster}

\item{path_raster_folder}{Character. Name a path folder to save raster images data. If  doesn't exist, a new directory is created}
}
\value{
Matrix with raster and pixels to create a RasterBrick with result
}
\description{
Update a RasterBrick with new values of pixel discovered from LUC Calculus formalism to create GeoTIFF files
}
\examples{
\dontrun{

rb_new <- lucC_save_raster_result(raster_obj = rb_sits, data_mtx = third_raster.df,
timeline = timeline, label = label, path_raster_folder = NULL)
rb_new

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
