% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_save_raster_GeoTIFF.R
\docType{data}
\name{lucC_save_rasterBrick_layers}
\alias{lucC_save_rasterBrick_layers}
\title{Save a RasterBrick by layers}
\usage{
lucC_save_rasterBrick_layers(path_name_GeoTIFF_Brick = NULL)
}
\arguments{
\item{path_name_GeoTIFF_Brick}{Character. Name file and path folder to SPLIT rasterBrick by layers}
}
\value{
A set of GeoTIFF separate images
}
\description{
Save a RasterBrick into individual layers files GeoTIFFs
}
\examples{
\dontrun{

lucC_save_raster_result(path_name_GeoTIFF_Brick =
"/home/inpe/Desktop/TesteIta/Mosaic_Raster_Splitted_.tif")

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
