% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_treat_columns.R
\docType{data}
\name{lucC_select_columns}
\alias{lucC_select_columns}
\title{Select columns of a data set}
\usage{
lucC_select_columns(data_mtx = NULL, name_columns = NULL)
}
\arguments{
\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{name_columns}{Character. Name of columns to remove from data set}
}
\value{
Matrix with without selected columns
}
\description{
Remove columns from a data_mtx
}
\examples{
\dontrun{

# remove columns
lucC_select_columns (data_mtx = dataset, name_columns = "2001-09-01")

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
