% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_standard_date_events.R
\docType{data}
\name{lucC_standard_date_events}
\alias{lucC_standard_date_events}
\title{lucC Apply TWDTW from SITS Format}
\usage{
lucC_standard_date_events(data_tb = NULL, 
month_year = "09", day_month = "01")
}
\arguments{
\item{data_tb}{Tibble. A tibble with values in sits format}

\item{month_year}{Character. A month of the year in two digits (01 to 12)}

\item{day_month}{Character. A day of the month in two digits (01 to 31). 
Depends the character calendar month.}
}
\value{
Tibble containing two new columns with start_date and end_date in a 
predefined standard range ensurer
}
\description{
Create two new columns in tibble with standard date in 
order to extract events
}
\examples{
\dontrun{

library(lucC)

data("example_TWDTW")
example_TWDTW

# alter start_date and end_date to a especific range in 
# order to extract events
example_1.tb <- example_TWDTW \%>\% 
  lucC_standard_date_events(data_tb = ., 
  month_year = "09", day_month = "01")

example_1.tb

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
