% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_load_data.R
\docType{data}
\name{lucC_starting_point}
\alias{lucC_starting_point}
\title{Format data as tibble format}
\usage{
lucC_starting_point()
}
\value{
Change the global default setting
}
\description{
Change the global default setting to not auto-convert to factors
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

}



}
\author{
Adeline M. Maciel
}
\keyword{datasets}
