% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_save_raster_GeoTIFF.R
\docType{data}
\name{lucC_toGeoTIFF}
\alias{lucC_toGeoTIFF}
\title{Save tibble as Raster in Folder}
\usage{
lucC_toGeoTIFF (data_tb = NULL, path_raster_folder = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{path_raster_folder}{Character. Name a path folder to save raster images data}
}
\value{
Images in geotiff format to open using SIG
}
\description{
Plot map ggplot2 for all events discovered in input data
}
\examples{
\dontrun{

# save rasters in folder
lucC_toGeoTIFF (example_TWDTW, "~/Desktop/raster")

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
