% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_create_blocks.R
\docType{data}
\name{lucC_create_blocks}
\alias{lucC_create_blocks}
\title{Create blocks from RasterBrick in accordance with number of cells}
\usage{
lucC_create_blocks (raster_obj = NULL, number_cells = 250)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{number_cells}{Numeric. A number of cells to split the raster in small parts. Default is 250.}
}
\value{
List of raster splitted by blocks
}
\description{
Provide a data.frame with main measures to resultd from LUC Calculus application such as: Area (km2),
Cumulative Sum (CumSum), Relative Frequency and Cumulative Relative Frequency
}
\examples{
\dontrun{

lucC_create_blocks(raster_obj = rb_sits2, number_cells = 250)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
