% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_create_blocks.R
\docType{data}
\name{lucC_merge_blocks}
\alias{lucC_merge_blocks}
\title{Create blocks from RasterBrick in accordance with number of cells}
\usage{
lucC_merge_blocks (path_open_GeoTIFFs = NULL, number_raster = 4,
pattern_name = NULL, is.rasterBrick = FALSE)
}
\arguments{
\item{path_open_GeoTIFFs}{Character. Name a path folder to OPEN raster images data.}

\item{number_raster}{Integer. Number of GeoTIFF files.}

\item{pattern_name}{Character. A pattern in name of GeoTIFF to mosaic them}

\item{is.rasterBrick}{Boolean. If TRUE GeoTIFF is a RasterBrick, FALSE is a single layer. Default is FALSE}
}
\value{
RasterBrick Mosaic.
}
\description{
Merge GeoTIFF splitted into parts. \url{https://stackoverflow.com/questions/29784829/r-raster-package-split-image-into-multiples}
}
\examples{
\dontrun{

lucC_merge_blocks(path_open_GeoTIFFs = NULL, number_raster = 4,
                  pattern_name = "MT_year_", is.rasterBrick = FALSE)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
