% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_holds_raster.R
\docType{data}
\name{lucC_pred_recur}
\alias{lucC_pred_recur}
\title{Predicate Allen Holds}
\usage{
lucC_pred_recur (raster_obj = NULL, raster_class = NULL,
time_interval1 = c("2001-01-01", "2001-01-01"),
time_interval2 = c("2003-01-01", "2004-01-01"),
label = NULL, timeline = NULL)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{raster_class}{Character. Name of class present in a row of the tibble, such as 'Forest' or other value}

\item{time_interval1}{Interval. A interval of time to verify if class is over or
not in character format. Given a tibble with values, will be asserts if that
class of locations holds during a time interval}

\item{time_interval2}{Interval. A interval of time to verify if class is over or
not in character format. Given a tibble with values, will be asserts if that
class of locations holds during a time interval}

\item{label}{Character. All labels of each value of pixel of classified raster}

\item{timeline}{Date. A list with all end_dates of classified raster}
}
\value{
Tibble with all events hold during a time interval
}
\description{
Provide a predicate of Allen's which asserts that a class holds
during a time interval. Return a tibble with value within defined interval
}
\examples{
\dontrun{

library(lucC)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
