% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations_extra.r
\docType{data}
\name{lucC_relation_follows}
\alias{lucC_relation_follows}
\title{Allen Relation Follows}
\usage{
lucC_relation_follows(first_raster = NULL, second_raster = NULL)
}
\arguments{
\item{first_raster}{matrix. An interval between two dates.}

\item{second_raster}{matrix. An interval between two dates.}
}
\value{
Data set with merge of two data sets
}
\description{
Provide an Allen's interval relation to classified raster data.
And return data set with raster data sets merged
}
\examples{
\dontrun{

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
