% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_data.R
\name{timeline_SVM_mt}
\alias{timeline_SVM_mt}
\alias{timeline_SVM_mt}
\title{Example CSV with a timeline}
\usage{
data(timeline_SVM_mt)
}
\description{
CSV file format of example of a timelines already classified using classification method.

Dataset of example of a timelines already classified using classification method.
}
